feat(toc): implement table of contents viewer
Adds the ability to view the Table of Contents (TOC) for a PDF file
within a Telescope picker.
A new command, `:PdfReader showToc`, is introduced to trigger the TOC
viewer for the current book.

Additionally, the list of recent books is now sorted to show the most
recently read first.